pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ensemble-genetic-algorithm"
version = "0.1.0"
description = "A project for ensemble methods using genetic algorithms."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Samora Hunter", email = "samorahunter@hotmail.co.uk" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

# Core dependencies required for the project to run
dependencies = [
    "deap",
    "eventlet",
    "fsspec",
    "fuzzysearch",
    "graphviz",
    "imbalanced-learn",
    "ipykernel",
    "ipywidgets",
    "matplotlib",
    "numpy<2",
    "pandas",
    "polars",
    "pydot",
    "pyimpetus",
    "scikit-learn",
    "scipy",
    "scoop",
    "seaborn",
    "session-info",
    "statsmodels",
    "tabulate",
    "torch==2.0.1",
    "torchmetrics",
    "xgboost",
    "pyarrow",
    "tqdm",
    "PyYAML",
]

[project.optional-dependencies]
# Dependencies for development, testing, and code quality
dev = [
    "pytest",
    "pre-commit",
    "nbformat", 
    "nbconvert", 
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
    "sphinx-autodoc-typehints",
]

# Dependencies for GPU support (e.g., CUDA-enabled libraries)
gpu = []

# 'all' combines dev and gpu dependencies
all = [
    "ensemble-genetic-algorithm[dev,gpu]",
]

[tool.setuptools]
# Explicitly tell setuptools which directory contains the package source code.
packages = ["ml_grid"]